+++ crates/release_plz/tests/all/changelog.rs
@@ -215,7 +215,7 @@ async fn can_generate_single_changelog_for_multiple_packages_in_pr() {
     let context = TestContext::new_workspace_with_packages(&[
         TestPackage::new("one")
             .with_type(PackageType::Bin)
-            .with_path_dependencies(vec![format!("../two")]),
+            .with_path_dependencies(vec!["../two".to_string()]),
         TestPackage::new("two").with_type(PackageType::Lib),
     ])
     .await;